+++ src/events/invite.h
@@ -1,6 +1,7 @@
 #ifndef GUARD_INVITE_H
 #define GUARD_INVITE_H
 
-void event_invite(struct irc_message_compo *compo);
+void event_inviting(struct irc_message_compo *);
+void event_invite(struct irc_message_compo *);
 
 #endif